refactor: move shared chown logic to uucore
